Added forEachIn, fillArray, and Dictionary to UTIL
These functions are all utilities that will be used in future pieces of
the project:
ForEachIn takes an object and runs a function on each of its
properties.
fillArray initializes or fills an array using a provide function of the
items index up to a certain length.
Dictionary is an object type that possesses an object whose only members
are defined explicitly and so can be accessed quickly and efficiently.