Renamed StaffingMatrix to Matrix and abstracted
Matrix is now composed of two arrays of Dictionaries that each store
identical values. Because javascript (sometimes) stores properties has
hashmaps, by using dictionaries, this should be more efficient for running
through entries to count them or decide whether or not they are set. This
should be usable for both staffing matrices, and possible the
ConflictMatrix.