Mainly documentation; added getXVals to Matrix
Documented the dictionary and matrix classes and also fixed up the
scriptDoc notation in Utils. Also added a function getXVals to Matrix
which gets a list of the keys in an x row of the matrix, essentially a
list of the y values for which the x value is defined.