Added necessary functionality of Staffer
Implemented a staff() method which places volunteers into the staffing
matrix based primarily on their own availability and the possibility of
conflicts. In order to do this, an AvailabilityIterator class was
necessary which keeps track of the position in both timeslot and person so
that staffing is done in the correct order and while preserving the proper
states. Also for this functionality, two new methods were added to UTIL:
greatestIndex and leastIndex, which given an array (and an optional array
which contains true/false values for whether or not the values in the
other array should be considered) finds the least (or greatest) value in
the array and returns the index of it. Also, some minor corrections to
documentation were made.